requirements.txt
arrow
asyncpg
aiopypixel
discord.py
python-dotenv
